rgw: move rgw_obj_key to rgw_basic_types.{h,cc}
